requirements.txt
aio-pika==6.7.1
aiohttp==3.7.2
aiopg==1.0.0
aiormq==3.3.1
async-timeout==3.0.1
asyncpg==0.21.0
attrs==20.3.0
certifi==2020.11.8
chardet==3.0.4
elasticsearch==7.10.0
idna==2.10
multidict==5.0.2
pamqp==2.3.0
peewee==3.14.0
peewee-async==0.7.1
pika==1.1.0
pip-autoremove==0.9.1
psycopg2-binary==2.8.6
python-dotenv==0.15.0
redis==3.5.3
requests==2.25.0
typing-extensions==3.7.4.3
urllib3==1.26.2
yarl==1.6.3
